Improve JobRepositoryTestUtils to mitigate OptimisticLockingFailureException while removing JobExecution
It's safe to remove stale version of JobExecution for testing purpose, this commit mitigate OptimisticLockingFailureException but cannot avoid since race condition may happen.

Fix GH-5161


(cherry picked from commit 1d9536cfbabf54b921f5a522beef62e5d0793a97)

+++ spring-batch-test/src/main/java/org/springframework/batch/test/JobRepositoryTestUtils.java
@@ -153,7 +153,12 @@ public class JobRepositoryTestUtils {
 	 * @param jobExecution the {@link JobExecution} to delete
 	 */
 	public void removeJobExecution(JobExecution jobExecution) {
-		this.jobRepository.deleteJobExecution(jobExecution);
+		// query latest version of JobExecution to avoid OptimisticLockingFailureException
+		jobExecution = this.jobRepository.getLastJobExecution(jobExecution.getJobInstance().getJobName(),
+				jobExecution.getJobParameters());
+		if (jobExecution != null) {
+			this.jobRepository.deleteJobExecution(jobExecution);
+		}
 	}
 
 	/**
@@ -170,7 +175,7 @@ public class JobRepositoryTestUtils {
 			while (!jobInstances.isEmpty()) {
 				for (JobInstance jobInstance : jobInstances) {
 					List<JobExecution> jobExecutions = this.jobRepository.findJobExecutions(jobInstance);
-					if (jobExecutions != null && !jobExecutions.isEmpty()) {
+					if (!jobExecutions.isEmpty()) {
 						removeJobExecutions(jobExecutions);
 					}
 				}
